fix netrules_access_check()
netrules_access_check() in aa.py checks if 
    type(netrules['rule'][family]) == dict
however this check always returns false (at least with py3, I didn't
test with py2).

This broken type check is the reason for
https://bugs.launchpad.net/apparmor/+bug/1380368
    aa-logprof doesn't propose abstractions for network rules
and 
https://bugs.launchpad.net/apparmor/+bug/1380367
    aa-logprof asks for already existing network rules
which are both fixed with this patch.


The type check is needed because netrules['rule'][family] can be
boolean True (for rules like "network inet,") - see line 2994.

The sock_type in .... .keys() check is there to ensure the hasher doesn't 
automagically add an empty sub-dict, which caused the regression in the 
first version of my patch.


Acked-by: Kshitij Gupta <kgupta8592@gmail.com>